increase clock speed to 168mhz, add fast/slow configs

+++ src/mcu/time.c
@@ -5,7 +5,7 @@
 static volatile uint32_t systicks = 0;
 
 void time_init(void) {
-  uint16_t cpu_ticks_per_ms = (sysclock_get_cpu_hz() / 1000);
+  uint32_t cpu_ticks_per_ms = (sysclock_get_cpu_hz() / 1000);
 
   SYSTICK->RVR = cpu_ticks_per_ms - 1; // set reload value for 1 ms
   SYSTICK->CVR = 0UL;                  // clear the value register
